+++ packages/e2e-tests/tests/smoke/order/mint-bid-cancel.test.ts
@@ -7,7 +7,7 @@ import type { OrderRequest } from "@rarible/sdk/src/types/order/common"
 import { toBigNumber } from "@rarible/types"
 import {
 	getEthereumWallet,
-	getEthereumWalletBuyer, getSolanaWallet,
+	getEthereumWalletBuyer, getFlowBuyerWallet, getFlowSellerWallet, getSolanaWallet,
 	getWalletAddressFull,
 } from "../../common/wallet"
 import { createSdk } from "../../common/create-sdk"
@@ -195,6 +195,32 @@ function suites(): {
 				}
 			},
 		},
+		{
+			blockchain: Blockchain.FLOW,
+			description: "NFT <=> FLOW_FT",
+			wallets: { seller: getFlowSellerWallet(), buyer: getFlowBuyerWallet() },
+			collectionId: testsConfig.variables.FLOW_RARIBLE_COLLECTION,
+			mintRequest: (creatorAddress: UnionAddress): MintRequest => {
+				return {
+					uri: "ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG5",
+					creators: [{
+						account: creatorAddress,
+						value: 10000,
+					}],
+					royalties: [],
+					lazyMint: false,
+					supply: 1,
+				}
+			},
+			currency: "FLOW_FT",
+			bidRequest: async (currency: RequestCurrency): Promise<OrderRequest> => {
+				return {
+					amount: 1,
+					price: "0.0001",
+					currency: currency,
+				}
+			},
+		},
 	]
 	return allBlockchains.filter(b => testsConfig.blockchain?.includes(b.blockchain))
 }
@@ -222,8 +248,8 @@ describe.each(suites())("$blockchain mint => bid => cancel", (suite) => {
 		const bidOrder = await bid(buyerSdk, buyerWallet, { itemId: nft.id }, bidRequest)
 
 		await getActivitiesByItem(buyerSdk, nft.id,
-			[ActivityType.MINT, ActivityType.BID],
-			[ActivityType.MINT, ActivityType.BID])
+			[ActivityType.BID],
+			[ActivityType.BID])
 
 		// Cancel order
 		await cancel(buyerSdk, buyerWallet, { orderId: bidOrder.id })