+++ packages/solana-sdk/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.3](https://github.com/rarible/sdk/compare/v0.9.2...v0.9.3) (2022-05-19)
+
+
+### Bug Fixes
+
+* set solana default commitment level ([#226](https://github.com/rarible/sdk/issues/226)) ([fdb7477](https://github.com/rarible/sdk/commit/fdb747747b5d7e05e46dcb3430fb894e3b9bc958))
+* solana process metadata ([#224](https://github.com/rarible/sdk/issues/224)) ([7e62a4c](https://github.com/rarible/sdk/commit/7e62a4cdb78bdd7fbe3f3656b4bbb2506201c061))
+
+
+
+
+
 # [0.9.0](https://github.com/rarible/sdk/compare/v0.8.52-hotfix5...v0.9.0) (2022-04-29)
 
 **Note:** Version bump only for package @rarible/solana-sdk